Made some singletons into classes, concurrency, fixes
I'm trying to get the wallpaper to render again. I am successfully parenting the window to the new WorkerW, and am rendering to a DC on the child instead of the WorkerW, but for some reason it is not appearing on screen. SPY++ reports that it is in the correct place. Removing WS_POPUP and WS_CHILD from the window will unparent the child and begin rendering successfully again, but this time in front of everything else. Unlike previous glitches, however, it can successfully be clicked away from like a regular window.